requirements.txt
langchain>=0.2.7,<0.3.0
langchain-postgres>=0.0.9,<0.1.0
sentence-transformers>=3.0.1,<4.0.0
langchain-huggingface>=0.0.3,<0.1.0
psycopg-binary>=3.2.1,<4.0.0
langchain-community>=0.2.14,<0.3.0
pdfplumber>=0.11.4,<0.12.0
openai>=1.43.0,<2.0.0
langchain-core
psycopg2-binary>=2.9.7,<3.0.0
requests==2.28.1
fastapi>=0.95.0,<1.0.0
uvicorn>=0.18.3,<0.19.0
python-dotenv>=1.0.0,<2.0.0
